Move AuthorizerError to error module
Moves the `AuthorizerError` struct from the
`splinter::network::connection_manager` module to the `error`
sub-module.
